requirements.txt
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
Unidecode==1.3.8
